+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## 4.5.2
 
 - [utils] fix: Decycling for objects to no produce an endless loop
+- [browser] fix: <unlabeled> event for unhandledRejection
+- [loader] fix: Handle unhandledRejection the same way as it would be thrown
 
 ## 4.5.1